flask/requirements.txt
uwsgi
flask
mysql-connector-python
